server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["nguyenddat <DatND.B23CC032@stu.ptit.edu.vn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.32.1"
fastapi = "^0.115.5"
fastapi-sqlalchemy = "^0.2.1"
apscheduler = "^3.11.0"
numpy = "^2.1.3"
pillow = "^11.0.0"
opencv-python = "^4.10.0.84"
pyjwt = "^2.10.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.10"
insightface = "^0.7.3"
onnxruntime = "^1.20.1"
pydantic = {extras = ["email"], version = "^2.10.1"}
python-jose = "^3.3.0"
python-multipart = "^0.0.17"
python-docx = "^1.1.2"
pandas = "^2.2.3"
qrcode = "^8.0"
pyzbar = "^0.1.9"
reportlab = "^4.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
